relocate api request methods
