+++ app/src/main/java/com/nudha/weatherapp/Activities/MainActivity.java
@@ -44,11 +44,11 @@ import retrofit2.Response;
 public class MainActivity extends AppCompatActivity {
     private RecyclerView.Adapter adapterHourly;
     private RecyclerView recyclerView;
-    private LocationUtils locationUtils;
+    private LocationUtils locationUtils;    //*******
     private TextView tempNow, highLowTemp,
             percipitation_now, wind_speed, uvIndx;
     private ImageView iconNow;
-    private SharedPreferences sharedPreferences;
+    private SharedPreferences sharedPreferences;    //*******
 
     private SimpleDateFormat sdf = new SimpleDateFormat("HH");
 
@@ -57,7 +57,7 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        sharedPreferences = getSharedPreferences("MyPrefs", MODE_PRIVATE);
+        sharedPreferences = getSharedPreferences("MyPrefs", MODE_PRIVATE); //*******
 
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
@@ -68,12 +68,6 @@ public class MainActivity extends AppCompatActivity {
         percipitation_now = findViewById(R.id.percipitation_now_TextView);
         wind_speed = findViewById(R.id.wind_speed_now_TextView);
         uvIndx = findViewById(R.id.uvIndx_now_TextView);
-
-
-        locationUtils = new LocationUtils(this);
-        locationUtils.requestLocation();
-
-
 
 
         setVariable();
@@ -143,13 +137,6 @@ public class MainActivity extends AppCompatActivity {
         recyclerView.setAdapter(adapterHourly);
     }
 
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        locationUtils.handlePermissionResult(requestCode, permissions, grantResults);
-
-    }
-
     public void setDataTime(){
         SimpleDateFormat dateFormat = new SimpleDateFormat("EEE MMM dd | HH:mm", Locale.ENGLISH);
         Date date = new Date();
@@ -159,88 +146,31 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void setWeatherNow(){
-        //https://api.meteomatics.com/2024-08-02T00:00:00ZPT1H/t_2m:C,t_max_2m_24h:C,t_min_2m_24h:C,precip_1h:mm,wind_speed_10m:ms,uv:idx,weather_symbol_1h:idx/50,10/json
-
-        //t_2:C - actual temperature
-        //t_max_2m_24h:C - max temperature
-        //t_min_2m_24h:C - min temperature
-        //precip_1h:mm - precipitation
-        //wind_speed_10m:ms - wind speed
-        //uv:idx - uv index
-        //weather_symbol_1h:idx - weather symbol
-
-        Log.d("MainActivity","Location: " + LocationPartRequest.getLocationCoordinates());
-        ApiService.getInstance().changeBaseUrl("https://api.meteomatics.com/");
-
-        String parameters = TempPartRequest.getTemp() + ","
-                + TempPartRequest.getTempStats("max24H") + ","
-                + TempPartRequest.getTempStats("min24H") + ","
-                + PrecipitationPartRequest.getPrecipitationPart("1h") + ","
-                + WindSpeedPartRequest.getWindSpeedPart() + ","
-                + "uv:idx" + "," + "weather_symbol_1h:idx" ;
-
-
-        ApiService.getInstance().getWeatherApi().getWeather(TimePartRequest.timeConvert("now"),
-                parameters, LocationPartRequest.getLocationCoordinates()).enqueue(new Callback<WeatherResponse>() {
-            @Override
-            public void onResponse(@NonNull Call<WeatherResponse> call, @NonNull Response<WeatherResponse> response) {
-                Log.d("MainActivity", "Response: " + response.code());
-                Log.d("MainActivity", "Response: " + response.message());
-                Log.d("MainActivity", "Call: " + call.request().url().toString());
-                if(response.isSuccessful()){
-                    WeatherResponse weatherResponse = response.body();
-                    Log.d("MainActivity", "Response: " + weatherResponse);
-                    if (weatherResponse != null && weatherResponse.getData() != null
-                            && !weatherResponse.getData().isEmpty()) {
-                        tempNow.setText(setData(weatherResponse, TempPartRequest.getTemp(), "tempNow"));
-
-                        Log.d("MainActivity", "Response body: " + weatherResponse.getData());
-                        String highTemp = setData(weatherResponse, TempPartRequest.getTempStats("max24H"), "highTemp");
-                        String lowTemp = setData(weatherResponse, TempPartRequest.getTempStats("min24H"), "lowTemp");
-                        String highLowTempString ="H: " + highTemp + "  L: " + lowTemp;
-
-                        highLowTemp.setText(highLowTempString);
-
-                        percipitation_now.setText(setData(weatherResponse, PrecipitationPartRequest.getPrecipitationPart("1h"), "percipitation_now") + " mm");
-
-                        wind_speed.setText(setData(weatherResponse, WindSpeedPartRequest.getWindSpeedPart(), "wind_speed") + " m/s");
-
-                        uvIndx.setText(setData(weatherResponse, "uv:idx", "uvIndx"));
-
-                    }
-                }
-            }
-
-            @Override
-            public void onFailure(Call<WeatherResponse> call, Throwable t) {
-                Log.e("MainActivity", "Error: " + t.getMessage());
-                Toast.makeText(MainActivity.this, "Error: " + t.getMessage(), Toast.LENGTH_SHORT).show();
-            }
-        });
-    }
-
-    private String setData(WeatherResponse weatherResponse, String parameter, String paramName) {
-        // Получаем данные из ответа
-        WeatherResponse.Data data = findParameter(weatherResponse.getData(), parameter);
-        WeatherResponse.Data.Coordinate coordinate = data.getCoordinates().get(0);
-        WeatherResponse.Data.Coordinate.DateValue dateValue = coordinate.getDates().get(0);
-
-        // Сохраняем данные в SharedPreferences
-        SharedPreferences.Editor editor = sharedPreferences.edit();
-        editor.putString(paramName, String.valueOf(dateValue.getValue()));
-        editor.apply();
-        Log.d("Saved param", String.valueOf(editor));
-
-        return String.valueOf(dateValue.getValue());
-    }
-
-    private WeatherResponse.Data findParameter(List<WeatherResponse.Data> data, String parameter) {
-        for (WeatherResponse.Data datum : data) {
-            if (datum.getParameter().equals(parameter)) {
-                return datum;
+        Bundle extras = getIntent().getExtras();
+        if (extras != null) {
+            for (String key : extras.keySet()) {
+                Object value = extras.get(key);
+                // Используйте ключ и значение как вам нужно
+                Log.d("NextActivity", "Key: " + key + " Value: " + value);
             }
         }
-        return null;
     }
+
+//    private String setData(WeatherResponse weatherResponse, String parameter, String paramName) {
+//        // Получаем данные из ответа
+//        WeatherResponse.Data data = findParameter(weatherResponse.getData(), parameter);
+//        WeatherResponse.Data.Coordinate coordinate = data.getCoordinates().get(0);
+//        WeatherResponse.Data.Coordinate.DateValue dateValue = coordinate.getDates().get(0);
+//
+//        // Сохраняем данные в SharedPreferences
+//        SharedPreferences.Editor editor = sharedPreferences.edit();
+//        editor.putString(paramName, String.valueOf(dateValue.getValue()));
+//        editor.apply();
+//        Log.d("Saved param", String.valueOf(editor));
+//
+//        return String.valueOf(dateValue.getValue());
+//    }
+
+
 
 }
